moves files to NeuralNetwork
