--- src/DiyNeuralNetwork/NeuralNetworkVis.js
@@ -1,7 +1,0 @@
-class NeuralNetworkVis {
-  constructor(){
-    this.temp = null;
-  }
-}
-
-export default NeuralNetworkVis;
